1) long is 64bits on 64bits systems, so be careful with random generator. 2) Make Point2i use int32_t rather than int. 3) Let fixedpoint class use fixint_t typedef to easily switch, and therefore adapt the action Pop/Push (that works!?)

+++ trunk/lib/wormux/include/WORMUX_random.h
@@ -48,9 +48,9 @@ public:
   Double GetDouble();
   Double GetDouble(Double max);
   Double GetDouble(Double min, Double max);
-  int GetInt(int min, int max);
-  uint GetUint(uint min, uint max);
-  long GetLong(long min, long max);
+  int32_t GetInt(int32_t min, int32_t max);
+  uint GetUint(uint32_t min, uint32_t max);
+  int32_t GetLong(int32_t min, int32_t max);
   Point2i GetPoint(const Rectanglei &rect);
   Point2i GetPoint(const Point2i &pt);
   int GetSign();
